+++ CHANGES.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - Better handling of terminal colors 
 - Send keynames to LS when updated, to record the modified stamp
+- Add cache control headers
 
 ##Version 0.1.7
 **May 17, 2014**